+++ web/testing/attachment-api/utilities.js
@@ -3,24 +3,24 @@ function loadKeyboards()
   var kmw=keyman;
 
   // The first keyboard added will be the default keyboard for touch devices.
-  // For faster loading, it may be best for the default keybaord to be 
+  // For faster loading, it may be best for the default keyboard to be
   // locally sourced.
-  kmw.addKeyboards({id:'us',name:'English',languages:{id:'eng',name:'English'},
+  kmw.addKeyboards({id:'us',name:'English',languages:{id:'en',name:'English'},
     filename:'../us-1.0.js'});
     
   // Add more keyboards to the language menu, by keyboard name,
-  // keyboard name and language code, or just the ISO 639 language code.  
-  kmw.addKeyboards('french','european2@swe','european2@nor','@heb');
+  // keyboard name and language code, or just the BCP-47 language code.
+  kmw.addKeyboards('french','european2@sv','european2@no','@he');
 
   // Add a keyboard by language name.  Note that the name must be spelled
-  // correctly, or the keyboard will not be found.  (Using ISO codes is
+  // correctly, or the keyboard will not be found.  (Using BCP-47 codes is
   // usually easier.)
   kmw.addKeyboardsForLanguage('Dzongkha');
 
   // Add a fully-specified, locally-sourced, keyboard with custom font  
   kmw.addKeyboards({id:'lao_2008_basic',name:'Lao Basic',
     languages:{
-    id:'lao',name:'Lao',region:'Asia',
+    id:'lo',name:'Lao',region:'Asia',
     font:{family:'LaoWeb',source:['../font/saysettha_web.ttf','../font/saysettha_web.woff','../font/saysettha_web.eot']}
     },
     filename:'../lao_2008_basic.js'