+++ web/unit_tests/cases/engine.js
@@ -58,7 +58,7 @@ var toEscapedSupplementaryPairString = function(code){
 var DEADKEY_TEST_1 = {
   id: 1,
   // Match condition for rule
-  rule: [{d: 1}],
+  rule: [{t:'d', d: 1}],
   // Start of context relative to cursor
   n: 1,
   ln: 1,
@@ -96,7 +96,7 @@ var DEADKEY_TEST_1 = {
 var DEADKEY_TEST_2 = {
   id: 2,
   // Match condition for rule
-  rule: ['a', {d: 0}, {d: 1}, 'b'],
+  rule: ['a', {t:'d', d: 0}, {t:'d', d: 1}, 'b'],
   // Start of context relative to cursor
   n: 5,
   ln: 4,
@@ -164,7 +164,7 @@ var DEADKEY_TEST_2 = {
 var DEADKEY_TEST_3 = {
   id: 3,
   // Match condition for rule
-  rule: [{d: 0}, 'a', {d: 0}, {d: 0}, 'b'],
+  rule: [{t:'d', d: 0}, 'a', {t:'d', d: 0}, {t:'d', d: 0}, 'b'],
   // Start of context relative to cursor
   n: 5,
   ln: 5,
@@ -252,7 +252,7 @@ var DEADKEY_TEST_3 = {
 var DEADKEY_TEST_4 = {
   id: 4,
   // Match condition for rule
-  rule: ['a', {d: 0}, {d: 0}, 'b', {d: 0}],
+  rule: ['a', {t:'d', d: 0}, {t:'d', d: 0}, 'b', {t:'d', d: 0}],
   // Start of context relative to cursor
   n: 6,
   ln: 5,
@@ -294,7 +294,7 @@ var DEADKEY_TEST_5 = {
 var DEADKEY_TEST_6 = {
   id: 6,
   // Match condition for rule
-  rule: [{d: 1}, {d: 2}, {d: 0}, {d: 1}, {d: 2}],
+  rule: [{t:'d', d: 1}, {t:'d', d: 2}, {t:'d', d: 0}, {t:'d', d: 1}, {t:'d', d: 2}],
   // Start of context relative to cursor
   n: 5,
   ln: 5,
@@ -315,7 +315,7 @@ var DEADKEY_TEST_6 = {
 var ANY_CONTEXT_TEST_1 = {
   id: 1,
   // Match condition for rule
-  rule: ['c', "a", "b", {c:3}, {c:2}],
+  rule: ['c', "a", "b", {t:'c', c:3}, {t:'c', c:2}],
   // Start of context relative to cursor
   n: 5,
   ln: 5,
@@ -355,7 +355,7 @@ var ANY_CONTEXT_TEST_1 = {
 var ANY_CONTEXT_TEST_2 = {
   id: 2,
   // Match condition for rule
-  rule: ['c', 'a', {a: "bc"}, {c:3}, 'a'],
+  rule: ['c', 'a', {t:'a', a: "bc"}, {t:'c', c:3}, 'a'],
   // Start of context relative to cursor
   n: 5,
   ln: 5,
@@ -395,7 +395,7 @@ var ANY_CONTEXT_TEST_2 = {
 var ANY_CONTEXT_TEST_3 = {
   id: 3,
   // Match condition for rule
-  rule: ['c', {a: "ac"}, {a: "bc"}, {c:3}, {c:2}],
+  rule: ['c', {t:'a', a: "ac"}, {t:'a', a: "bc"}, {t:'c', c:3}, {t:'c', c:2}],
   // Start of context relative to cursor
   n: 5,
   ln: 5,
@@ -445,7 +445,7 @@ var ANY_CONTEXT_TEST_3 = {
 var ANY_INDEX_TEST_1 = {
   id: 1,
   // Match condition for rule
-  rule: ['c', 'a', {a: "bc"}, {i:{s:"bc", o:3}}, 'a'],
+  rule: ['c', 'a', {t:'a', a: "bc"}, {t:'i', i:{s:"bc", o:3}}, 'a'],
   // Start of context relative to cursor
   n: 5,
   ln: 5,
@@ -485,7 +485,7 @@ var ANY_INDEX_TEST_1 = {
 var ANY_INDEX_TEST_2 = {
   id: 2,
   // Match condition for rule
-  rule: ['c', {a:"ab"}, {i: {s:"bc", o:2}}, {i:{s:"bc", o:2}}, {i:{s:"ab", o:2}}],
+  rule: ['c', {t:'a', a:"ab"}, {t:'i', i: {s:"bc", o:2}}, {t:'i', i:{s:"bc", o:2}}, {t:'i', i:{s:"ab", o:2}}],
   // Start of context relative to cursor
   n: 5,
   ln: 5,
@@ -525,7 +525,7 @@ var ANY_INDEX_TEST_2 = {
 var ANY_INDEX_TEST_3 = {
   id: 3,
   // Match condition for rule
-  rule: ['c', {a:"ab"}, {a:"bc"}, {i:{s:"bc", o:3}}, {i:{s:"ab", o:2}}],
+  rule: ['c', {t:'a', a:"ab"}, {t:'a', a:"bc"}, {t:'i', i:{s:"bc", o:3}}, {t:'i', i:{s:"ab", o:2}}],
   // Start of context relative to cursor
   n: 5,
   ln: 5,