+++ web/history.md
@@ -25,6 +25,7 @@
 * Fixed styling of "Special" keys on touch layouts
 * Added automated testing for KeymanWeb builds. (#350)
 * Fixed bugs in the handling of deadkeys. (#281)
+* Change from ISO 639-3 language codes to BCP-47 language codes
 
 ## 2017-07-10 2.0.473 stable
 * 2.0 stable release build.